fix: Exporting models properly. and changing README.
