+++ lib/src/models.dart
@@ -125,3 +125,36 @@ class CIDInfo {
   }
   */
 }
+
+
+/// Not all multibase spec base types are supported by this library.
+/// We support the `default` ones from the official repo in https://github.com/multiformats/multibase/blob/master/multibase.csv.
+enum Multibase {
+  /// hexadecimal
+  base16(code: 'f', name: "base16"),
+
+  /// hexadecimal
+  base16upper(code: 'F', name: "base16upper"),
+
+  /// rfc4648 case-insensitive - no padding
+  base32(code: 'b', name: "base32"),
+
+  /// rfc4648 case-insensitive - no padding
+  base32upper(code: 'B', name: "base32upper"),
+
+  /// base58 bitcoin
+  base58btc(code: 'z', name: "base58btc"),
+
+  /// rfc4648 no padding
+  base64(code: 'm', name: "base64");
+
+  /// Initialize a new multibase variant
+  const Multibase({
+    required final String code,
+    required final String name,
+  })  : baseCode = code,
+        baseName = name;
+
+  final String baseCode;
+  final String baseName;
+}
