+++ test/dart_multibase_test.dart
@@ -1,4 +1,5 @@
 import 'package:dart_cid/src/decode_cid.dart';
+import 'package:dart_cid/src/models.dart';
 import 'package:dart_cid/src/multibase.dart';
 import 'package:test/test.dart';
 import 'dart:typed_data';
